changed from home page to blogs page (#17)
